Move smart pointer aliases to Tbx/Memory/Refs.h
