+++ Engine/Include/Tbx/Plugins/Plugin.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Tbx/DllExport.h"
+#include "Tbx/Memory/Refs.h"
 #include <memory>
-#include "Tbx/Memory/Refs/Refs.h"
 
 namespace Tbx
 {
