+++ Examples/3dDemo/App/Demo.h
@@ -1,11 +1,11 @@
 #pragma once
+#include "Tbx/Memory/Refs.h"
 #include <Tbx/App/App.h>
 #include <Tbx/TSS/Toy.h>
 #include <Tbx/TSS/Stage.h>
 #include <Tbx/Graphics/Material.h>
 #include <Tbx/Layers/Layer.h>
 #include <Tbx/Plugins/Plugin.h>
-#include "Tbx/Memory/Refs/Refs.h"
 
 class Demo : public Tbx::Layer, public Tbx::Plugin
 {
